separate logic into modules
