+++ lib/repor/calculator.rb
@@ -1,0 +1,2 @@
+require 'repor/calculator/base'
+require 'repor/calculator/ratio'
