+++ lib/repor/serializer.rb
@@ -1,0 +1,6 @@
+require 'repor/serializer/base'
+require 'repor/serializer/table'
+require 'repor/serializer/csv'
+require 'repor/serializer/form_field'
+require 'repor/serializer/highcharts'
+require 'repor/serializer/hash_table'
