+++ lib/repor/tracker.rb
@@ -1,0 +1,3 @@
+require 'repor/tracker/base'
+require 'repor/tracker/count'
+require 'repor/tracker/ratio'
