+++ spec/repor/serializer/hash_table_spec.rb
