+++ spec/repor/serializer/highcharts_spec.rb
