+++ spec/repor/serializer/table_spec.rb
